scripts/pipeline/requirements.txt
feedparser>=6.0.10
requests>=2.31.0
beautifulsoup4>=4.12.0
google-genai>=0.3.0
python-frontmatter>=1.0.1
